refactor : domain, utils, view 분리
